+++ networks/ethereum.config.ts
@@ -54,6 +54,7 @@ export const ethereum: NetworkUserConfig = {
     chainlink: {
       deployLocal: false,
       deleteOldJobs: false,
+      cleanLocalFolder: false,
       nodeFunds: '1',
       gasLimit: undefined,
       ethWsUrl: process.env.ETHEREUM_WS_URI,
@@ -108,7 +109,6 @@ export const ethereum: NetworkUserConfig = {
     messengers: [
       {
         contract: CONTRACT_NAMES.SEMessenger,
-        specificationPath: `${appRoot.path}/contracts/messengers/${USE_CASES.STAKING_EFFICIENCY}/${CONTRACT_NAMES.SEMessenger}.json`,
         useCaseName: USE_CASES.STAKING_EFFICIENCY,
         externalAdapterUrl: process.env.STAKING_EFFICIENCY_INDEXER_URI,
       },