+++ services/external-adapter/env-validation.ts
@@ -1,22 +1,19 @@
-import hardhat from 'hardhat';
+import * as hardhat from 'hardhat';
 
 const Joi = require('joi');
 require('dotenv').config();
 
-const networksObject = Object.keys(hardhat.config.networks)
-  .filter((network) => !['hardhat', 'localhost'].includes(network))
-  .reduce(
-    (r, network) => ({
-      ...r,
-      [`${network.toUpperCase()}_URI`]: Joi.string().uri().required(),
-    }),
-    {}
-  );
-
-const schema = Joi.object({
-  IPFS_URI: Joi.string().uri().required(),
-  ...networksObject,
-}).unknown();
+const schema = Joi.object(
+  Object.keys(hardhat.config.networks)
+    .filter((network) => !['hardhat', 'localhost'].includes(network))
+    .reduce(
+      (r, network) => ({
+        ...r,
+        [`${network.toUpperCase()}_URI`]: Joi.string().uri().required(),
+      }),
+      {}
+    )
+).unknown();
 
 const { error } = schema.validate(process.env);