+++ services/external-adapter/index.ts
@@ -12,9 +12,7 @@ type SLAData = {
   serviceURL: string;
   serviceAddress: string;
   serviceTicker: string;
-
   periodType: number;
-  networkName: string;
   messengerAddress: string;
 };
 
@@ -29,12 +27,11 @@ async function getSLAData(address): Promise<SLAData> {
   const web3 = new Web3(web3Uri);
   const slaContract = new web3.eth.Contract(SLAABI, address);
   const ipfsCID = await slaContract.methods.ipfsHash().call();
-  console.log(`SLA IPFS url: ${process.env.IPFS_URI}/ipfs/${ipfsCID}`);
+  console.log(`SLA IPFS url: https://ipfs.io/ipfs/${ipfsCID}`);
   const periodType = await slaContract.methods.periodType().call();
-  const networkName = await slaContract.methods.extraData(0).call();
   const messengerAddress = await slaContract.methods.messengerAddress().call();
-  const { data } = await axios.get(`${process.env.IPFS_URI}/ipfs/${ipfsCID}`);
-  return { ...data, periodType, networkName, messengerAddress };
+  const { data } = await axios.get(`https://ipfs.io/ipfs/${ipfsCID}`);
+  return { ...data, periodType, messengerAddress };
 }
 
 async function getSLI(requestData: RequestData) {