+++ subtasks.ts
@@ -35,6 +35,7 @@ import {
   PERIOD_STATUS,
   PERIOD_TYPE,
   TOKEN_NAMES,
+  USE_CASES,
 } from './constants';
 import {
   bootstrapStrings,
@@ -46,7 +47,7 @@ import {
 import axios from 'axios';
 import { HardhatRuntimeEnvironment } from 'hardhat/types';
 import { formatBytes32String } from 'ethers/lib/utils';
-import { networks } from './hardhat.config';
+import { networks } from './networks';
 
 const prettier = require('prettier');
 const appRoot = require('app-root-path');
@@ -170,8 +171,19 @@ subtask(SUB_TASK_NAMES.SETUP_DOCKER_COMPOSE, undefined).setAction(
     const { stacktical } = network.config;
     const linkToken = await get(CONTRACT_NAMES.LinkToken);
 
+    if (stacktical.chainlink.cleanLocalFolder) {
+      const folders = fs
+        .readdirSync(`${appRoot.path}/services/chainlink-nodes/`)
+        .filter((folder) => new RegExp(`${network.name}`).test(folder));
+      for (let folder of folders) {
+        fs.rmdirSync(`${appRoot.path}/services/chainlink-nodes/${folder}`, {
+          recursive: true,
+        });
+      }
+    }
     for (let node of stacktical.chainlink.nodesConfiguration) {
       const nodeName = network.name + '-' + node.name;
+
       const fileContents = fs.readFileSync(
         `${appRoot.path}/services/docker-compose.yaml`,
         'utf8'
@@ -192,21 +204,6 @@ subtask(SUB_TASK_NAMES.SETUP_DOCKER_COMPOSE, undefined).setAction(
               }`;
             case /CHAINLINK_PORT/.test(envVariable):
               return `CHAINLINK_PORT=${node.restApiPort}`;
-            // case /ETH_GAS_PRICE_DEFAULT/.test(envVariable):
-            //   if (network.config.gasPrice) {
-            //     return `ETH_GAS_PRICE_DEFAULT=${network.config.gasPrice}`;
-            //   }
-            //   return envVariable;
-            // case /ETH_MAX_GAS_PRICE_WEI/.test(envVariable):
-            //   if (network.config.gasPrice) {
-            //     return `ETH_MAX_GAS_PRICE_WEI=${network.config.gasPrice}`;
-            //   }
-            //   return envVariable;
-            // case /ETH_MIN_GAS_PRICE_WEI/.test(envVariable):
-            //   if (network.config.gasPrice) {
-            //     return `ETH_MIN_GAS_PRICE_WEI=${network.config.gasPrice}`;
-            //   }
-            //   return envVariable;
             default:
               return envVariable;
           }
@@ -731,10 +728,8 @@ subtask(SUB_TASK_NAMES.BOOTSTRAP_MESSENGER_REGISTRY, undefined).setAction(
     for (let messenger of messengers) {
       console.log('Registering ' + messenger.contract + ' on the SLARegistry');
       const messengerArtifact = await get(messenger.contract);
-
-      const messengerSpec = JSON.parse(
-        fs.readFileSync(messenger.specificationPath)
-      );
+      const specificationPath = `${appRoot.path}/contracts/messengers/${messenger.useCaseName}/use-case-spec.json`;
+      const messengerSpec = JSON.parse(fs.readFileSync(specificationPath));
       const updatedSpec = {
         ...messengerSpec,
         timestamp: new Date().toISOString(),
@@ -826,9 +821,8 @@ subtask(SUB_TASK_NAMES.DEPLOY_MESSENGER, undefined).setAction(
       consola.info(
         'Registering ' + messenger.contract + ' in MessengerRegistry'
       );
-      const messengerSpec = JSON.parse(
-        fs.readFileSync(messenger.specificationPath)
-      );
+      const specificationPath = `${appRoot.path}/contracts/messengers/${messenger.useCaseName}/use-case-spec.json`;
+      const messengerSpec = JSON.parse(fs.readFileSync(specificationPath));
       const updatedSpec = {
         ...messengerSpec,
         timestamp: new Date().toISOString(),
@@ -1013,10 +1007,8 @@ subtask(SUB_TASK_NAMES.DEPLOY_SLA, undefined).setAction(
       const stakeAmount =
         Number(initialTokenSupply) / initialTokenSupplyDivisor;
       const stakeAmountTimesWei = (times) => toWei(String(stakeAmount * times));
-      const semessengerArtifact = await get(CONTRACT_NAMES.SEMessenger);
-      const seMessenger = await SEMessenger__factory.connect(
-        semessengerArtifact.address,
-        signer
+      const messenger: IMessenger = await ethers.getContract(
+        config.messengerContract
       );
       const ipfsHash = await getIPFSHash(serviceMetadata);
       const stakeRegistryArtifact = await get(CONTRACT_NAMES.StakeRegistry);
@@ -1053,10 +1045,10 @@ subtask(SUB_TASK_NAMES.DEPLOY_SLA, undefined).setAction(
 
       console.log('Starting process 2: Deploy SLA');
       tx = await slaRegistry.createSLA(
-        sloValue,
+        sloValue * Number(await messenger.messengerPrecision()),
         sloType,
         whitelisted,
-        seMessenger.address,
+        messenger.address,
         periodType,
         initialPeriodId,
         finalPeriodId,
@@ -1103,10 +1095,13 @@ subtask(SUB_TASK_NAMES.DEPLOY_SLA, undefined).setAction(
       console.log(
         `Starting process 3.2: notDeployer: ${fromWei(notDeployerStake)} DSLA`
       );
-      await dslaToken.connect(await ethers.getSigner(notDeployer));
-      tx = await dslaToken.approve(sla.address, notDeployerStake);
+      tx = await dslaToken
+        .connect(await ethers.getSigner(notDeployer))
+        .approve(sla.address, notDeployerStake);
       await tx.wait();
-      tx = await sla.stakeTokens(notDeployerStake, dslaToken.address);
+      tx = await sla
+        .connect(await ethers.getSigner(notDeployer))
+        .stakeTokens(notDeployerStake, dslaToken.address);
       await tx.wait();
       printSeparator();
     }
@@ -1375,7 +1370,7 @@ subtask(SUB_TASK_NAMES.FULFILL_SLI, undefined).setAction(
     // const signer = await ethers.getSigner(deployer);
     // const seMessenger = await SEMessenger__factory.connect(
     //   (
-    //     await get(CONTRACT_NAMES.SEMessenger)
+    //     await get(CONTRACT_NAMES.BaseMessenger)
     //   ).address,
     //   signer
     // );