+++ types.ts
@@ -1,4 +1,11 @@
-import {CONTRACT_NAMES, PERIOD_TYPE, SLO_TYPE, TOKEN_NAMES,} from './constants';
+import {
+  CONTRACT_NAMES,
+  PERIOD_TYPE,
+  SLO_TYPE,
+  TOKEN_NAMES,
+  USE_CASES,
+} from './constants';
+import { ContractFactory } from 'ethers';
 
 export type StackticalConfiguration = {
   checkPastPeriods: boolean;
@@ -34,6 +41,7 @@ export type DeployedContractAddresses = {
 export type ChainlinkConfiguration = {
   deployLocal: boolean;
   deleteOldJobs: boolean;
+  cleanLocalFolder: boolean;
   nodeFunds: string;
   gasLimit?: string;
   ethWsUrl: string;
@@ -84,9 +92,8 @@ export type TokenAllowance = {
 
 export type MessengerConfiguration = {
   contract: CONTRACT_NAMES;
-  specificationPath: string;
+  useCaseName: USE_CASES;
   externalAdapterUrl: string;
-  useCaseName: string;
 };
 
 export type DeploySLAConfiguration = {
@@ -100,12 +107,13 @@ export type DeploySLAConfiguration = {
   };
   sloValue: number;
   sloType: SLO_TYPE;
-  whitelisted: boolean;
   periodType: PERIOD_TYPE;
   initialPeriodId: number;
   finalPeriodId: number;
-  extraData: Array<string>;
   leverage: number;
+  messengerContract: CONTRACT_NAMES;
+  whitelisted: boolean;
+  extraData: Array<string>;
   initialTokenSupply: string;
   initialTokenSupplyDivisor: number;
   deployerStakeTimes: number;